airbyte-integrations/connectors/source-amocrm-incremental/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-amocrm-incremental"
description = "Source implementation for amocrm-incremental."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/amocrm-incremental"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_amocrm_incremental" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = ">3.9,<3.12"
airbyte-cdk = "^0"

[tool.poetry.scripts]
source-amocrm-incremental = "source_amocrm_incremental.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"
flake8 = "^7.0.0"
